requirements.txt
PyExifTool